requirements.txt
Django==2.0.1
django-autoslug==1.9.4.dev0
django-storages==1.6.5
factory-boy==2.10.0
Faker==0.8.11
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
text-unidecode==1.2
